Moved falstad stuff, commented
I added a lot of comments, moved the Falstad.kt file into the parsers
namespace, and broke out all of the components to make it easier to see
the core of the Falstad importer.

A bunch of Javadoc comments, but Falstad.kt still needs a bunch of
comments.
