+++ core/src/main/kotlin/org/eln2/parsers/falstad/components/generic/OutputProbe.kt
@@ -1,0 +1,32 @@
+package org.eln2.parsers.falstad.components.generic
+
+import org.eln2.parsers.falstad.CCData
+import org.eln2.parsers.falstad.IComponentConstructor
+import org.eln2.parsers.falstad.PinRef
+import org.eln2.parsers.falstad.PosSet
+import org.eln2.sim.electrical.mna.component.Resistor
+
+/**
+ * Output Probe
+ *
+ * Falstad's Output Probe - shows the voltage of a particular node in the circuit.
+ */
+class OutputProbe: IComponentConstructor {
+	companion object {
+		val HIGH_IMPEDANCE = Double.POSITIVE_INFINITY
+	}
+
+	override fun construct(ccd: CCData) {
+		val r = Resistor()
+		r.resistance = HIGH_IMPEDANCE
+		ccd.circuit.add(r)
+		r.connect(1, ccd.circuit.ground)
+
+		val pp = (ccd.falstad.getPin(ccd.pos).representative as PosSet)
+		val pr = PinRef(r, 0)
+		ccd.falstad.addPinRef(pp, pr)
+		ccd.falstad.addOutput(pr)
+
+		ccd.pins = 1
+	}
+}
